pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "svcctl"
version = "0.2.0"
description = "Multi-app service manager"
requires-python = ">=3.11"
dependencies = [
    "click>=8",
    "pyyaml>=6",
    "textual>=0.50",
]

[project.scripts]
svcctl = "svcctl.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["svcctl*"]
